Security and data type changes for dev local coordinator mode
